Added generated files to gitignore (#2)
Everything in `.vscode`, `.pio` and `.DS_STORE` is locally generated on
your device. These directories should not/cannot be shared between users
because they will be generated differently depending on the user's
platform and settings (i.e. Mac, Windows, Linux, IDE in use)

I've added these three items to .gitignore so that git will not track
any changes in them so they will not be pushed to github. Note that two
of the items are postfixed with a `/` to denote that they are
directories and we should not track any file within them.